fix: Outlines No Corners selection box floating point precision.

+++ src/main/java/fi/dy/masa/malilib/render/RenderUtils.java
@@ -74,6 +74,9 @@ import fi.dy.masa.malilib.interfaces.IGuiRendererInvoker;
 import fi.dy.masa.malilib.mixin.client.IMixinMinecraft;
 import fi.dy.masa.malilib.mixin.gui.IMixinGuiRenderer;
 import fi.dy.masa.malilib.mixin.render.IMixinGameRenderer;
+import fi.dy.masa.malilib.mixin.render.IMixinAbstractTexture;
+import fi.dy.masa.malilib.mixin.render.IMixinGuiGraphics;
+import fi.dy.masa.malilib.mixin.render.IMixinGuiRenderer;
 import fi.dy.masa.malilib.render.element.*;
 import fi.dy.masa.malilib.render.on_demand.SelectionBoxRenderer;
 import fi.dy.masa.malilib.render.on_demand.TextPlateRenderer;
@@ -2902,19 +2905,19 @@ public class RenderUtils
         final double dy = cameraPos.y;
         final double dz = cameraPos.z;
 
-        final float dxMin = (float) (-dx - expand);
-        final float dyMin = (float) (-dy - expand);
-        final float dzMin = (float) (-dz - expand);
-        final float dxMax = (float) (-dx + expand);
-        final float dyMax = (float) (-dy + expand);
-        final float dzMax = (float) (-dz + expand);
+		final double dxMin = -dx - expand;
+		final double dyMin = -dy - expand;
+		final double dzMin = -dz - expand;
+		final double dxMax = -dx + expand;
+		final double dyMax = -dy + expand;
+		final double dzMax = -dz + expand;
 
-        final float minX = xMin + dxMin;
-        final float minY = yMin + dyMin;
-        final float minZ = zMin + dzMin;
-        final float maxX = xMax + dxMax;
-        final float maxY = yMax + dyMax;
-        final float maxZ = zMax + dzMax;
+		final float minX = (float) (xMin + dxMin);
+		final float minY = (float) (yMin + dyMin);
+		final float minZ = (float) (zMin + dzMin);
+		final float maxX = (float) (xMax + dxMax);
+		final float maxY = (float) (yMax + dyMax);
+		final float maxZ = (float) (zMax + dzMax);
 
         int start, end;
 
@@ -2928,8 +2931,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(start + dxMin, minY, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
-            buffer.addVertex(end + dxMax, minY, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (start + dxMin), minY, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (end + dxMax), minY, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMax && pos1.getZ() == zMin) || (pos2.getX() == xMin && pos2.getY() == yMax && pos2.getZ() == zMin) ? xMin + 1 : xMin;
@@ -2937,8 +2940,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(start + dxMin, maxY + 1, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
-            buffer.addVertex(end + dxMax, maxY + 1, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (start + dxMin), maxY + 1, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (end + dxMax), maxY + 1, minZ).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMin && pos1.getZ() == zMax) || (pos2.getX() == xMin && pos2.getY() == yMin && pos2.getZ() == zMax) ? xMin + 1 : xMin;
@@ -2946,8 +2949,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(start + dxMin, minY, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
-            buffer.addVertex(end + dxMax, minY, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (start + dxMin), minY, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (end + dxMax), minY, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMax && pos1.getZ() == zMax) || (pos2.getX() == xMin && pos2.getY() == yMax && pos2.getZ() == zMax) ? xMin + 1 : xMin;
@@ -2955,8 +2958,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(start + dxMin, maxY + 1, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
-            buffer.addVertex(end + dxMax, maxY + 1, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (start + dxMin), maxY + 1, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
+            buffer.addVertex((float) (end + dxMax), maxY + 1, maxZ + 1).setColor(colorX.r, colorX.g, colorX.b, colorX.a).setLineWidth(lineWidth);
         }
 
         // Edges along the Y-axis
@@ -2965,8 +2968,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(minX, start + dyMin, minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
-            buffer.addVertex(minX, end + dyMax, minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, (float) (start + dyMin), minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, (float) (end + dyMax), minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMax && pos1.getY() == yMin && pos1.getZ() == zMin) || (pos2.getX() == xMax && pos2.getY() == yMin && pos2.getZ() == zMin) ? yMin + 1 : yMin;
@@ -2974,8 +2977,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(maxX + 1, start + dyMin, minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
-            buffer.addVertex(maxX + 1, end + dyMax, minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, (float) (start + dyMin), minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, (float) (end + dyMax), minZ).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMin && pos1.getZ() == zMax) || (pos2.getX() == xMin && pos2.getY() == yMin && pos2.getZ() == zMax) ? yMin + 1 : yMin;
@@ -2983,8 +2986,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(minX, start + dyMin, maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
-            buffer.addVertex(minX, end + dyMax, maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, (float) (start + dyMin), maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, (float) (end + dyMax), maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMax && pos1.getY() == yMin && pos1.getZ() == zMax) || (pos2.getX() == xMax && pos2.getY() == yMin && pos2.getZ() == zMax) ? yMin + 1 : yMin;
@@ -2992,8 +2995,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(maxX + 1, start + dyMin, maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
-            buffer.addVertex(maxX + 1, end + dyMax, maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, (float) (start + dyMin), maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, (float) (end + dyMax), maxZ + 1).setColor(colorY.r, colorY.g, colorY.b, colorY.a).setLineWidth(lineWidth);
         }
 
         // Edges along the Z-axis
@@ -3002,8 +3005,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(minX, minY, start + dzMin).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
-            buffer.addVertex(minX, minY, end + dzMax).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, minY, (float) (start + dzMin)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, minY, (float) (end + dzMax)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMax && pos1.getY() == yMin && pos1.getZ() == zMin) || (pos2.getX() == xMax && pos2.getY() == yMin && pos2.getZ() == zMin) ? zMin + 1 : zMin;
@@ -3011,8 +3014,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(maxX + 1, minY, start + dzMin).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
-            buffer.addVertex(maxX + 1, minY, end + dzMax).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, minY, (float) (start + dzMin)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, minY, (float) (end + dzMax)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMin && pos1.getY() == yMax && pos1.getZ() == zMin) || (pos2.getX() == xMin && pos2.getY() == yMax && pos2.getZ() == zMin) ? zMin + 1 : zMin;
@@ -3020,8 +3023,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(minX, maxY + 1, start + dzMin).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
-            buffer.addVertex(minX, maxY + 1, end + dzMax).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, maxY + 1, (float) (start + dzMin)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(minX, maxY + 1, (float) (end + dzMax)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
         }
 
         start = (pos1.getX() == xMax && pos1.getY() == yMax && pos1.getZ() == zMin) || (pos2.getX() == xMax && pos2.getY() == yMax && pos2.getZ() == zMin) ? zMin + 1 : zMin;
@@ -3029,8 +3032,8 @@ public class RenderUtils
 
         if (end > start)
         {
-            buffer.addVertex(maxX + 1, maxY + 1, start + dzMin).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
-            buffer.addVertex(maxX + 1, maxY + 1, end + dzMax).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, maxY + 1, (float) (start + dzMin)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
+            buffer.addVertex(maxX + 1, maxY + 1, (float) (end + dzMax)).setColor(colorZ.r, colorZ.g, colorZ.b, colorZ.a).setLineWidth(lineWidth);
         }
 
         try
@@ -3039,7 +3042,6 @@ public class RenderUtils
 
             if (meshData != null)
             {
-//                ctx.lineWidth(lineWidth);
                 ctx.draw(meshData, false, true);
                 meshData.close();
             }
